requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0
tensorboard>=2.12.0
jupyter>=1.0.0
notebook>=6.5.0
pandas>=2.0.0
seaborn>=0.12.0
albumentations>=1.3.0
